DRT-4381 Start a test to compare PHP optimiser output with R

+++ shared/src/main/scala/spatutorial/shared/Api.scala
@@ -1,6 +1,6 @@
 package spatutorial.shared
 
-case class CrunchResult(recommendedDesks: Seq[Double], waitTimes: Seq[Double])
+case class CrunchResult(recommendedDesks: Seq[Int], waitTimes: Seq[Int])
 
 trait Api {
   def welcomeMsg(name: String): String
